Seperates slice what has all about status by concerns
Added writeSlice and WriteSlice.test.js
