+++ src/state/slice.test.js
@@ -2,17 +2,17 @@ import thunk from 'redux-thunk';
 import configureStore from 'redux-mock-store';
 
 import reducer, {
-  changeRadioChecked,
+  // changeRadioChecked,
   changeInputFieldValue,
   setInputFieldsError,
-  increaseWritePageIndex,
-  decreaseWritePageIndex,
-  flipPreviewPostcard,
+  // increaseWritePageIndex,
+  // decreaseWritePageIndex,
+  // flipPreviewPostcard,
   flipPostcard,
   setPostcardFront,
   // setEntrance,
-  setWriteCompleteValues,
-  resetPostcardInputFields,
+  // setWriteCompleteValues,
+  // resetPostcardInputFields,
   setPostcard,
   // admitPostcardAccess,
   setResponseError,
@@ -24,8 +24,8 @@ import reducer, {
   setMovingPage,
 
   // loadEntrance,
-  sendPhoto,
-  sendPostcard,
+  // sendPhoto,
+  // sendPostcard,
   // checkValidPostcard,
   loadPostcard,
   loadPostcards,
@@ -35,7 +35,7 @@ import reducer, {
 import {
   // postCheckValidPostcard,
   // fetchEntrance,
-  postPostcard,
+  // postPostcard,
   fetchPostcard,
   postExpire,
 } from '../services/api';
@@ -103,7 +103,7 @@ describe('reducer', () => {
         },
       };
       const initialState = {
-        writePageIndex: 0,
+        // writePageIndex: 0,
         inputFields: initialInputFields,
         // entrance: {
         //   sender: '테스트',
@@ -133,21 +133,21 @@ describe('reducer', () => {
     });
   });
 
-  describe('changeRadioChecked', () => {
-    it('changes isPrivate in write variable', () => {
-      const initialState = {
-        inputFields: {
-          write: {
-            isPrivate: false,
-          },
-        },
-      };
+  // describe('changeRadioChecked', () => {
+  //   it('changes isPrivate in write variable', () => {
+  //     const initialState = {
+  //       inputFields: {
+  //         write: {
+  //           isPrivate: false,
+  //         },
+  //       },
+  //     };
 
-      const state = reducer(initialState, changeRadioChecked({ value: true }));
+  //     const state = reducer(initialState, changeRadioChecked({ value: true }));
 
-      expect(state.inputFields.write.isPrivate.value).toBe(true);
-    });
-  });
+  //     expect(state.inputFields.write.isPrivate.value).toBe(true);
+  //   });
+  // });
 
   describe('changeInputFieldValue', () => {
     it('changes InputFieldValue', () => {
@@ -199,47 +199,47 @@ describe('reducer', () => {
     });
   });
 
-  describe('increaseWritePageIndex', () => {
-    it('increases writePageIndex + 1', () => {
-      const initialState = {
-        writePageIndex: 0,
-      };
+  // describe('increaseWritePageIndex', () => {
+  //   it('increases writePageIndex + 1', () => {
+  //     const initialState = {
+  //       writePageIndex: 0,
+  //     };
 
-      const state = reducer(initialState, increaseWritePageIndex());
+  //     const state = reducer(initialState, increaseWritePageIndex());
 
-      expect(state.writePageIndex).toBe(1);
-    });
-  });
+  //     expect(state.writePageIndex).toBe(1);
+  //   });
+  // });
 
-  describe('decreaseWritePageIndex', () => {
-    it('decreases writePageIndex -1', () => {
-      const initialState = {
-        writePageIndex: 1,
-      };
+  // describe('decreaseWritePageIndex', () => {
+  //   it('decreases writePageIndex -1', () => {
+  //     const initialState = {
+  //       writePageIndex: 1,
+  //     };
 
-      const state = reducer(initialState, decreaseWritePageIndex());
+  //     const state = reducer(initialState, decreaseWritePageIndex());
 
-      expect(state.writePageIndex).toBe(0);
-    });
-  });
+  //     expect(state.writePageIndex).toBe(0);
+  //   });
+  // });
 
-  describe('flipPreviewPostcard', () => {
-    it('change isFrontPage in preview inputField', () => {
-      const initialState = {
-        inputFields: {
-          write: {
-            preview: {
-              isFrontPage: false,
-            },
-          },
-        },
-      };
+  // describe('flipPreviewPostcard', () => {
+  //   it('change isFrontPage in preview inputField', () => {
+  //     const initialState = {
+  //       inputFields: {
+  //         write: {
+  //           preview: {
+  //             isFrontPage: false,
+  //           },
+  //         },
+  //       },
+  //     };
 
-      const state = reducer(initialState, flipPreviewPostcard());
+  //     const state = reducer(initialState, flipPreviewPostcard());
 
-      expect(state.inputFields.write.preview.isFrontPage).toBe(true);
-    });
-  });
+  //     expect(state.inputFields.write.preview.isFrontPage).toBe(true);
+  //   });
+  // });
 
   describe('flipPostcard', () => {
     it('change isFrontPage in postcard', () => {
@@ -285,126 +285,126 @@ describe('reducer', () => {
   //   });
   // });
 
-  describe('setWriteCompleteValues', () => {
-    it('set WriteCompleteValues in inputFields', () => {
-      const page = 'write';
-      const type = 'complete';
+  // describe('setWriteCompleteValues', () => {
+  //   it('set WriteCompleteValues in inputFields', () => {
+  //     const page = 'write';
+  //     const type = 'complete';
 
-      const url = 'url';
-      const secretMessage = 'secretMessage';
+  //     const url = 'url';
+  //     const secretMessage = 'secretMessage';
 
-      const initialState = {
-        inputFields: {
-          [page]: {
-            [type]: {
-              url: '',
-              secretMessage: '',
-            },
-          },
-        },
-      };
+  //     const initialState = {
+  //       inputFields: {
+  //         [page]: {
+  //           [type]: {
+  //             url: '',
+  //             secretMessage: '',
+  //           },
+  //         },
+  //       },
+  //     };
 
-      const state = reducer(initialState, setWriteCompleteValues({
-        url,
-        secretMessage,
-      }));
+  //     const state = reducer(initialState, setWriteCompleteValues({
+  //       url,
+  //       secretMessage,
+  //     }));
 
-      expect(state.inputFields[page][type].url).toBe(url);
-      expect(state.inputFields[page][type].secretMessage).toBe(secretMessage);
-    });
-  });
+  //     expect(state.inputFields[page][type].url).toBe(url);
+  //     expect(state.inputFields[page][type].secretMessage).toBe(secretMessage);
+  //   });
+  // });
 
-  describe('resetPostcardInputFields', () => {
-    it('reset resetPostcardInputFields in inputFields', () => {
-      const inputFields = {
-        entrance: {
-          value: '1',
-          error: 'false',
-        },
-        write: {
-          isPrivate: true,
-          secretMessage: {
-            value: 'secretMessage',
-            error: false,
-          },
-          sender: {
-            value: 'sender',
-            error: false,
-          },
-          receiver: {
-            value: 'receiver',
-            error: false,
-          },
-          contents: {
-            value: 'contents',
-            error: '',
-          },
-          photo: {
-            value: 'photo',
-            error: false,
-          },
-          photoMessage: {
-            value: 'photoMessage',
-            error: false,
-          },
-          preview: {
-            isFrontPage: true,
-          },
-          complete: {
-            key: 'complete',
-            secretMessage: '',
-          },
-        },
-      };
+  // describe('resetPostcardInputFields', () => {
+  //   it('reset resetPostcardInputFields in inputFields', () => {
+  //     const inputFields = {
+  //       entrance: {
+  //         value: '1',
+  //         error: 'false',
+  //       },
+  //       write: {
+  //         isPrivate: true,
+  //         secretMessage: {
+  //           value: 'secretMessage',
+  //           error: false,
+  //         },
+  //         sender: {
+  //           value: 'sender',
+  //           error: false,
+  //         },
+  //         receiver: {
+  //           value: 'receiver',
+  //           error: false,
+  //         },
+  //         contents: {
+  //           value: 'contents',
+  //           error: '',
+  //         },
+  //         photo: {
+  //           value: 'photo',
+  //           error: false,
+  //         },
+  //         photoMessage: {
+  //           value: 'photoMessage',
+  //           error: false,
+  //         },
+  //         preview: {
+  //           isFrontPage: true,
+  //         },
+  //         complete: {
+  //           key: 'complete',
+  //           secretMessage: '',
+  //         },
+  //       },
+  //     };
 
-      const initialState = { inputFields };
+  //     const initialState = { inputFields };
 
-      const state = reducer(initialState, resetPostcardInputFields());
+  //     const state = reducer(initialState, resetPostcardInputFields());
 
-      expect(state.inputFields).toEqual({
-        entrance: {
-          secretMessage: {
-            value: '',
-            error: false,
-          },
-        },
-        write: {
-          isPrivate: true,
-          secretMessage: {
-            value: '',
-            error: false,
-          },
-          sender: {
-            value: '',
-            error: false,
-          },
-          receiver: {
-            value: '',
-            error: false,
-          },
-          contents: {
-            value: '',
-            error: '',
-          },
-          photo: {
-            value: '',
-            error: false,
-          },
-          photoMessage: {
-            value: '',
-            error: false,
-          },
-          preview: {
-            isFrontPage: true,
-          },
-          complete: {
-            key: '',
-            secretMessage: '',
-          },
-        },
-      });
-    });
-  });
+  //     expect(state.inputFields).toEqual({
+  //       entrance: {
+  //         secretMessage: {
+  //           value: '',
+  //           error: false,
+  //         },
+  //       },
+  //       write: {
+  //         isPrivate: true,
+  //         secretMessage: {
+  //           value: '',
+  //           error: false,
+  //         },
+  //         sender: {
+  //           value: '',
+  //           error: false,
+  //         },
+  //         receiver: {
+  //           value: '',
+  //           error: false,
+  //         },
+  //         contents: {
+  //           value: '',
+  //           error: '',
+  //         },
+  //         photo: {
+  //           value: '',
+  //           error: false,
+  //         },
+  //         photoMessage: {
+  //           value: '',
+  //           error: false,
+  //         },
+  //         preview: {
+  //           isFrontPage: true,
+  //         },
+  //         complete: {
+  //           key: '',
+  //           secretMessage: '',
+  //         },
+  //       },
+  //     });
+  //   });
+  // });
 
   describe('setPostcard', () => {
     it('set postcard variables', () => {
@@ -625,87 +625,87 @@ describe('reducer', () => {
   //   });
   // });
 
-  describe('sendPhoto', () => {
-    beforeEach(() => {
-      store = mockStore({});
-    });
+  // describe('sendPhoto', () => {
+  //   beforeEach(() => {
+  //     store = mockStore({});
+  //   });
 
-    it('runs changeInputFieldValue', async () => {
-      const fileName = 'test';
-      await store.dispatch(sendPhoto({ file: new File([''], fileName) }));
+  //   it('runs changeInputFieldValue', async () => {
+  //     const fileName = 'test';
+  //     await store.dispatch(sendPhoto({ file: new File([''], fileName) }));
 
-      const actions = store.getActions();
+  //     const actions = store.getActions();
 
-      expect(actions[0]).toEqual(changeInputFieldValue({
-        page: 'write',
-        type: 'photo',
-        value: fileName,
-      }));
-    });
-  });
+  //     expect(actions[0]).toEqual(changeInputFieldValue({
+  //       page: 'write',
+  //       type: 'photo',
+  //       value: fileName,
+  //     }));
+  //   });
+  // });
 
-  describe('sendPostcard', () => {
-    beforeEach(() => {
-      store = mockStore({});
+  // describe('sendPostcard', () => {
+  //   beforeEach(() => {
+  //     store = mockStore({});
 
-      postPostcard.mockImplementation(() => Promise.resolve(given.response));
-    });
+  //     postPostcard.mockImplementation(() => Promise.resolve(given.response));
+  //   });
 
-    const postcard = {
-      key: 'test',
-      sender: 'sender',
-      receiver: 'receiver',
-      contents: 'contents',
-      photo: 'photo',
-      photoMessage: 'photoMessage',
-      secretMessage: 'secretMessage',
-      isPrivate: 'isPrivate',
-    };
-    const onClickNext = jest.fn();
+  //   const postcard = {
+  //     key: 'test',
+  //     sender: 'sender',
+  //     receiver: 'receiver',
+  //     contents: 'contents',
+  //     photo: 'photo',
+  //     photoMessage: 'photoMessage',
+  //     secretMessage: 'secretMessage',
+  //     isPrivate: 'isPrivate',
+  //   };
+  //   const onClickNext = jest.fn();
 
-    context('when response has error', () => {
-      it('runs setMovingPage', async () => {
-        given('response', () => ({
-          error: responseError,
-        }));
+  //   context('when response has error', () => {
+  //     it('runs setMovingPage', async () => {
+  //       given('response', () => ({
+  //         error: responseError,
+  //       }));
 
-        await store.dispatch(sendPostcard({
-          postcard,
-          onClickNext,
-        }));
+  //       await store.dispatch(sendPostcard({
+  //         postcard,
+  //         onClickNext,
+  //       }));
 
-        const actions = store.getActions();
+  //       const actions = store.getActions();
 
-        expect(actions[0]).toEqual(setResponseError(responseError));
-      });
-    });
+  //       expect(actions[0]).toEqual(setResponseError(responseError));
+  //     });
+  //   });
 
-    context('when response does not have error', () => {
-      it('runs changeInputFieldValue', async () => {
-        const url = 'url';
-        const secretMessage = 'secretMessage';
+  //   context('when response does not have error', () => {
+  //     it('runs changeInputFieldValue', async () => {
+  //       const url = 'url';
+  //       const secretMessage = 'secretMessage';
 
-        given('response', () => ({
-          data: {
-            url,
-            secretMessage,
-          },
-        }));
+  //       given('response', () => ({
+  //         data: {
+  //           url,
+  //           secretMessage,
+  //         },
+  //       }));
 
-        await store.dispatch(sendPostcard({
-          postcard,
-          onClickNext,
-        }));
+  //       await store.dispatch(sendPostcard({
+  //         postcard,
+  //         onClickNext,
+  //       }));
 
-        const actions = store.getActions();
+  //       const actions = store.getActions();
 
-        expect(actions[0]).toEqual(setWriteCompleteValues({
-          url,
-          secretMessage,
-        }));
-      });
-    });
-  });
+  //       expect(actions[0]).toEqual(setWriteCompleteValues({
+  //         url,
+  //         secretMessage,
+  //       }));
+  //     });
+  //   });
+  // });
 
   // describe('checkValidPostcard', () => {
   //   beforeEach(() => {
